Handle NOTICEs in isFromChannel and isFromUser
I don't know what these methods were originally intended to do as it is
not clear whether the word `message` in the function documentation
refers to PRIVMSGs or IRC server messages in general.

I personally use these two methods to figure out if a IRC messages was
sent from a user or channel and if so map it to the channel specified in
`event.Params[0]`. For this use case it would be nice if NOTICEs would
be treated the same way as they are automatically mapped to the correct
channel/user as well then.