Fix #63
Logic race condition, if the decode goroutine can run before the caller
can the channel returned might not have a receiver yet and then the
select/default construct will drop the event and close the channel.

The caller then ends up with a zero-value decodedEvent where
decodedEvent.err and decodedEvent.event are both nil which passes the
err check at line 434 and might enter line 440 where decodedEvent.event
is then used as-is with no additional check for it being nil.